modificação em cenario de controlador
