+++ test/cucumber/steps/CollectionPointsSteps.groovy
@@ -1,11 +1,11 @@
 package cucumber.steps
 
-import br.ufpe.cin.ines.ress.Address
+
 import br.ufpe.cin.ines.ress.PickupRequest
-import br.ufpe.cin.ines.ress.Role
+import br.ufpe.cin.ines.ress.SignUpController
 import br.ufpe.cin.ines.ress.User
-import br.ufpe.cin.ines.ress.UserRole
 import br.ufpe.cin.ines.ress.residuecollector.CollectorDashboardController
+import br.ufpe.cin.ines.ress.residuegenerator.GeneratorDashboardController
 import pages.CollectionPointsPage
 import pages.CollectorDashboardPage
 import pages.HomePage
@@ -40,10 +40,12 @@ And(~/^os restaurantes de login "([^"]*)" e "([^"]*)" possuem coletas pendentes$
     assert user != null
     assert PickupRequest.findByGenerator(user)
 }
+
 When(~/^eu solicito a visualização das coletas$/) { ->
     at MapsPage
     page.clickCollectionPoints()
 }
+
 Then(~/^eu vejo a localização dos restaurantes de login "([^"]*)" e "([^"]*)" no mapa da UFPE$/) { String arg1, String arg2 ->
     at CollectionPointsPage
     page.hasmap()
@@ -61,31 +63,15 @@ Then(~/^eu vejo uma mensagem sinalizando que não há restaurantes com coletas p
 
 Given(~/^o local de login "([^"]*)" possui coletas pendentes$/) { String arg ->
 
-    def generatorRole = Role.findByAuthority('ROLE_GENERATOR') ?: new Role(authority: 'ROLE_GENERATOR').save(failOnError: true)
-    def generator = User.findByUsername(arg) ?: new User(username: arg,
-            password: 'pass',
-            name: 'Restaurante Universitário - RU',
-            email: arg+ '@gmail.com',
-            address: new Address(street: 'Av. Reitor Joaquim Amazonas', cep: '50740-540', city: 'Recife', state: 'Pernambuco', streetNumber: '22', neighborhood: 'Cidade Universitária'),
-            enabled: true).save(failOnError: true)
-    if(!generator.authorities.contains(generatorRole)){
-        UserRole.create(generator, generatorRole, true)
-    }
+    def signUpController = new SignUpController()
+    signUpController.createDefaultGeneratorUser(arg)
 
-    def pickupRequest = new PickupRequest()
+    def generatorController  = new GeneratorDashboardController()
+    generatorController.saveDefaultPickUp(arg)
 
-    pickupRequest.generator = User.findByUsername(arg)
-    pickupRequest.date = new Date()
-    pickupRequest.status = false
-    pickupRequest.residueAmount = 40
-    pickupRequest.collector = User.findByUsername('admin')
-    pickupRequest.save()
+    def pickups = PickupRequest.findAllByStatus(false).findAll {it -> it.generator.username == arg}
 
-    def coletas = PickupRequest.findAllByStatus(false);
-    def generators = coletas.collect {it -> it.generator.username}
-    def nome = generators.findAll{it -> it == arg}
-
-    assert !nome.isEmpty()
+    assert !pickups.isEmpty()
 }
 
 def address
@@ -95,8 +81,8 @@ When(~/^eu solicito os enderecos dos locais com coletas pendentes$/) { ->
 }
 
 Then(~/^o sistema retorna os enderecos do locais de login "([^"]*)"$/) { String arg ->
-    def usernames = address.collect{at -> at.user.username}
-    assert usernames.contains(arg)
+
+    assert address.collect{at -> at.user.username}.contains(arg)
     PickupRequest.executeUpdate('delete from PickupRequest')
 }
